refactor: split config flow schema helpers
Move config flow schema construction and entity lookup helpers into a new
config_flow_schema module so _get_schema can be decomposed without pushing
config_flow.py over the aislop file-size threshold.

Keep schema field ordering and Required/Optional markers covered by a
focused test, and retarget config flow entity lookup mocks to the new
module.

Refs #772
